+++ packages/core/src/main/cryptography/AssetCryptography/interfaces.ts
@@ -1,0 +1,33 @@
+/*
+ * Wire
+ * Copyright (C) 2021 Wire Swiss GmbH
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see http://www.gnu.org/licenses/.
+ *
+ */
+
+export interface Crypto {
+  digest(algo: string, cipherText: Buffer | Uint8Array): Promise<Buffer>;
+
+  decrypt(cipherText: Buffer | Uint8Array, keyBytes: Buffer): Promise<Buffer>;
+
+  getRandomValues(size: number): Buffer;
+
+  encrypt(
+    plainText: Buffer | Uint8Array,
+    keyBytes: Buffer,
+    initializationVector: Buffer,
+    algorithm: string,
+  ): Promise<{key: Buffer; cipher: Buffer}>;
+}
